% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hdr.R
\name{FindShift}
\alias{FindShift}
\title{Find Shift function}
\usage{
FindShift(rho, X, A, b)
}
\arguments{
\item{rho}{Fraction of X that fall in the resulting domain}

\item{X}{A matrix. Each column is a sample from a (possibly constrained) multivariate Gaussian.}

\item{A}{A matrix that together with b defines a set of constraints such that x is in the constrained domain if all(A\%*\%x + b >0 ). Number of rows is dimension of space in which samples are obtained. number of columns is number of constraints.}

\item{b}{A vector that together with A defines a set of constraints such that x is in the constrained domain if all(A\%*\%x + b >0 ). Number of elements is number of constraints.}
}
\value{
A list is returned containing the following elements:
\item{gamma_scalar}{Scalar. Shift such that rho*N samples lie in the domain.}
\item{rho_hat}{Scalar. True fraction in domain after shift by gamma_scalar. Can deviate from rho.}
\item{inside_inds}{Indices of the columns in the constraint after shift by gamma_scalar.}
}
\description{
Find a shift such that a fraction rho of X fall into the resulting domain
}
