% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hdr.R
\name{HDR_algo}
\alias{HDR_algo}
\title{The Holmes-Diaconis-Ross algorithm applied to linearly constrained Gaussians (Algorithm 1 of Gessner et al.).}
\usage{
HDR_algo(A, b, gamma_vec, N, nskip = 0)
}
\arguments{
\item{A}{A matrix that together with b defines a set of constraints such that x is in the constrained domain if all(A\%*\%x + b >0 ). Number of rows is dimension of space in which samples are obtained. number of columns is number of constraints.}

\item{b}{A vector that together with A defines a set of constraints such that x is in the constrained domain if all(A\%*\%x + b >0 ). Number of elements is number of constraints.}

\item{gamma_vec}{Vector of shifts defining nested domains that contain the constrained domain defined by A and b.}

\item{N}{Number of samples to draw from the constrained domain.}
}
\value{
Approximation of log of probability (multi-dimensional integral) for the constrained domain.
}
\description{
The Holmes-Diaconis-Ross algorithm applied to linearly constrained Gaussians (Algorithm 1 of Gessner et al.).
}
