% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hdr.R
\name{LinESS}
\alias{LinESS}
\title{Elliptical slice sampling for a linearly constrained standard normal distribution (Algorithm 2 of Gessner et al.)}
\usage{
LinESS(A, b, N, x0, nskip = 0)
}
\arguments{
\item{A}{A matrix that together with b defines a set of constraints such that x is in the constrained domain if all(A\%*\%x + b >0 ). Number of rows is dimension of space in which samples are obtained. number of columns is number of constraints.}

\item{b}{A vector that together with A defines a set of constraints such that x is in the constrained domain if all(A\%*\%x + b >0 ). Number of elements is number of constraints.}

\item{N}{Number of samples to draw from the constrained domain.}

\item{x0}{Initial value. Must be a vector that lies inside the constrained domain.}
}
\value{
A matrix with number of columns equal to the number of samples, N, and number of rows equal to the the dimension of the space in which samples are obtained.
}
\description{
Elliptical slice sampling for a linearly constrained standard normal distribution (Algorithm 2 of Gessner et al.).
}
