% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrappersfastfunctions.R
\name{SubsetSimFast}
\alias{SubsetSimFast}
\title{Subset simulation for linear constraints (Algorithm 3 of Gessner et al. 2020) implemented in C++}
\usage{
SubsetSimFast(A, b, N, rho = 0.5, nskip = 0)
}
\arguments{
\item{A}{A matrix that together with b defines a set of constraints such that x is in the constrained domain if all(A\%*\%x + b >0 ). Number of rows is dimension of space in which samples are obtained. number of columns is number of constraints.}

\item{b}{A vector that together with A defines a set of constraints such that x is in the constrained domain if all(A\%*\%x + b >0 ). Number of elements is number of constraints.}

\item{N}{Number of samples to draw from the constrained domain.}

\item{rho}{Fraction of X that fall in the resulting domain}

\item{nskip}{Number of samples to skip in order to get more independent samples}
}
\value{
A list is returned containing the following elements:
\item{logZ}{Scalar. Approximation of log of probability (multi-dimensional integral) for the constrained domain. Biased estimate.}
\item{shift_seq}{Vector of shifts defining nested domains that contain the constrained domain defined by A and b.}
}
\description{
Subset simulation for linear constraints (Algorithm 3 of Gessner et al. 2020).
}
